+++ src/main/java/chess/controller/RestControllerAdvice.java
@@ -1,0 +1,14 @@
+package chess.controller;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+
+@org.springframework.web.bind.annotation.RestControllerAdvice
+public class RestControllerAdvice {
+
+    @ExceptionHandler({IllegalArgumentException.class, IllegalStateException.class})
+    private ResponseEntity<String> handleException(final RuntimeException exception) {
+        return new ResponseEntity<>(exception.getMessage(), HttpStatus.BAD_REQUEST);
+    }
+}
